Add files via upload
The primary purpose of this project is to create a machine learning model that can generate realistic and innovative fashion designs. By leveraging the power of GANs, which consist of a Generator and a Discriminator network, the project aims to produce high-quality fashion images that can mimic human creativity. The Generator creates new designs, while the Discriminator evaluates their authenticity, helping the model improve over time. This iterative process ensures that the fashion designs generated are not only visually appealing but also align with industry standards.